CORE-7943 Cascade up the integration test initilization logic
